service/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
# langchain==0.0.35
git+https://github.com/BenderV/langchain@master#egg=langchain
MarkupSafe==2.1.1
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
peewee==3.15.4
psycopg2==2.9.5
pydantic==1.10.2
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
text-unidecode==1.3
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
transformers==4.24.0